Added support for all and range to now for summary
Added support for `-s all` and `-s <from>:now`.  This necessitated that
the `log` extraction function use `from` and `to` instead of a regex, as
this was problematic when matching against all possible instances from
the beginning of time. The `expand_date_range` and `match_date_range`
functions are no longer needed, so they've been removed.