Supports --kill/--reset/-x on dropbox
Dropbox does not sync timestamp changes, so the normal behaviour of just
touching the `TT_FILE` does not work.  For Dropbox, we now write a
comment to the file that states when it was reset and how many minutes
were abandoned.  This not only triggers the timestamp sync, it also
self-documents in case you need to find out how much time was lost.